requirements.txt
asgiref==3.4.1
bcrypt==3.2.0
certifi==2021.10.8
cffi==1.15.0
cryptography==35.0.0
Django==3.2.8
django-cors-headers==3.10.0
pycparser==2.20
PyJWT==2.3.0
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
